Rename classes for clarity
